// FrontBookRentals/src/app/api-client/model/bookResponseDtoICollectionGenericResponse.ts
/**
 * Book Rentals API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BookResponseDto } from './bookResponseDto';


export interface BookResponseDtoICollectionGenericResponse { 
    success: boolean;
    message: string;
    data?: Array<BookResponseDto> | null;
}