// FrontBookRentals/src/app/api-client/model/orderByBookResponseDtoICollectionGenericResponse.ts
/**
 * Book Rentals API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderByBookResponseDto } from './orderByBookResponseDto';


export interface OrderByBookResponseDtoICollectionGenericResponse { 
    success: boolean;
    message: string;
    data?: Array<OrderByBookResponseDto> | null;
}