// FrontBookRentals/src/app/api-client/model/orderByClientResponseDtoICollectionGenericResponse.ts
/**
 * Book Rentals API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderByClientResponseDto } from './orderByClientResponseDto';


export interface OrderByClientResponseDtoICollectionGenericResponse { 
    success: boolean;
    message: string;
    data?: Array<OrderByClientResponseDto> | null;
}