// FrontBookRentals/src/app/api-client/model/orderResponseDtoICollectionGenericResponse.ts
/**
 * Book Rentals API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderResponseDto } from './orderResponseDto';


export interface OrderResponseDtoICollectionGenericResponse { 
    success: boolean;
    message: string;
    data?: Array<OrderResponseDto> | null;
}